First commit 2022-02-25 End Code

+++ src/main/java/com/manjunatha/zooplus/orderdetails/model/dto/info/OrderPaymentInfoDto.java
@@ -14,9 +14,9 @@ public class OrderPaymentInfoDto {
 	
 	private Long paymentId;
 	
-	private int customerId;
+	private Long customerId;
 	
-	private int orderId;
+	private Long orderId;
 	
 	private Date paymentDate;
